Disable dss-notify-dev on deploy (related to #108)
Sets the concurrent execution limit of dss-notify-dev to zero on
deploy, while increasing it to 1 when tests are being run. Per #108,
there is a need to remove the v1 notifier (on the basis of cost) but
removing it from the codebase entirely is prohibitive because of its
integration with the codebase; in particular, it is used in the
ElasticSearch code that is set to be removed at some point in the
future.

As such, it makes sense to effectively 'suspend' execution of
dss-notify-dev as a band-aid solution.